refactor(utils/navigation): do not use imports from '/pages/*' [YTFRONT-3814]
