+++ packages/ui/src/ui/components/ColumnCell/ColumnCell.scss
@@ -1,0 +1,22 @@
+.yt-column-cell {
+    position: relative;
+    padding: var(--data-table-cell-vertical-padding) var(--data-table-cell-horizontal-padding);
+    padding-right: calc(var(--data-table-cell-horizontal-padding) + 20px);
+
+    &__control-button-wrapper {
+        opacity: 0.3;
+
+        position: absolute;
+        top: 0;
+        right: 0;
+        padding: 0;
+
+        &:hover {
+            opacity: 0.5;
+        }
+    }
+
+    &__copy-value-tooltip {
+        white-space: nowrap;
+    }
+}
