+++ packages/ui/src/ui/pages/navigation/Navigation/ContentViewer/helpers/index.ts
@@ -1,3 +1,3 @@
 export {default as getComponentByContentType} from './getComponentByContentType';
 export {default as getComponentByMode} from './getComponentByMode';
-export {default as checkContentIsSupported} from './checkContentIsSupported';
+export {checkContentIsSupported, itemNavigationAllowed} from './checkContentIsSupported';
