+++ packages/ui/src/ui/utils/job/tabs/statistics.ts
@@ -1,6 +1,6 @@
 import ypath from '../../../common/thor/ypath';
-import Job from '../../../pages/operations/OperationDetail/tabs/Jobs/job-selector';
-import {JobStatistic, RawJob} from '../../../types/operations/job';
+import type Job from '../../../pages/operations/OperationDetail/tabs/Jobs/job-selector';
+import type {JobStatistic, RawJob} from '../../../types/operations/job';
 
 export function prepareStatistics(job: Job | RawJob): JobStatistic {
     // @ts-ignore
